feat: 메일 모듈화, README.md 추가
