+++ withins_server/api/build.gradle.kts
@@ -19,7 +19,6 @@ val jwtVersion: String by project
 
 dependencies {
     implementation("org.springframework.boot:spring-boot-starter-web")
-    implementation("org.springframework.boot:spring-boot-starter-mail")
     implementation("org.springframework.boot:spring-boot-starter-security")
     implementation("org.springframework.boot:spring-boot-starter-oauth2-client")
     implementation("org.springframework.boot:spring-boot-starter-oauth2-resource-server")
@@ -27,6 +26,7 @@ dependencies {
     implementation(project(":support:monitoring"))
     implementation(project(":support:logging"))
     implementation(project(":crawl-batch"))
+    implementation(project(":mail"))
 
     testImplementation("io.rest-assured:rest-assured:${restAssuredVersion}")
     testImplementation("org.springframework.boot:spring-boot-starter-data-jpa")
