--- withins_server/api/src/main/java/com/withins/api/mail/config/MailConfig.java
@@ -1,39 +1,0 @@
-package com.withins.api.mail.config;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.mail.javamail.JavaMailSenderImpl;
-
-import java.util.Properties;
-
-@Configuration
-@RequiredArgsConstructor
-public class MailConfig {
-
-    @Value("${spring.mail.host}")
-    private String host;
-    @Value("${spring.mail.port}")
-    private String port;
-    @Value("${spring.mail.username}")
-    private String username;
-    @Value("${spring.mail.password}")
-    private String password;
-
-    private final Properties mailProperties;
-
-    @Bean
-    JavaMailSender javaMailSender() {
-        JavaMailSenderImpl javaMailSender = new JavaMailSenderImpl();
-        javaMailSender.setHost(host);
-        javaMailSender.setPort(Integer.parseInt(port));
-        javaMailSender.setUsername(username);
-        javaMailSender.setPassword(password);
-        javaMailSender.setDefaultEncoding("UTF-8");
-
-        javaMailSender.setJavaMailProperties(mailProperties);
-        return javaMailSender;
-    }
-}
