--- withins_server/api/src/main/java/com/withins/api/mail/exceptions/MailExceptionHandler.java
@@ -1,31 +1,0 @@
-package com.withins.api.mail.exceptions;
-
-import com.withins.api.exception.GlobalExceptionHandler;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-import java.time.LocalDateTime;
-
-@RestControllerAdvice
-public class MailExceptionHandler {
-
-    @ExceptionHandler(EmailExternalException.class)
-    public ResponseEntity<GlobalExceptionHandler.ErrorResponse> handleMailExternalException(EmailExternalException e) {
-        GlobalExceptionHandler.ErrorResponse errorResponse = new GlobalExceptionHandler.ErrorResponse(
-            HttpStatus.BAD_REQUEST.value(),
-            e.getMessage(),
-            LocalDateTime.now());
-        return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
-    }
-
-    @ExceptionHandler(EmailInternalException.class)
-    public ResponseEntity<GlobalExceptionHandler.ErrorResponse> handleMailInternalException(EmailInternalException e) {
-        GlobalExceptionHandler.ErrorResponse errorResponse = new GlobalExceptionHandler.ErrorResponse(
-            HttpStatus.INTERNAL_SERVER_ERROR.value(),
-            e.getMessage(),
-            LocalDateTime.now());
-        return new ResponseEntity<>(errorResponse, HttpStatus.INTERNAL_SERVER_ERROR);
-    }
-}
