--- withins_server/api/src/main/java/com/withins/api/mail/exceptions/SMTPBadCredentialsException.java
@@ -1,8 +1,0 @@
-package com.withins.api.mail.exceptions;
-
-public class SMTPBadCredentialsException extends EmailInternalException {
-
-    public SMTPBadCredentialsException() {
-        super("SMTP 아이디 또는 비밀번호가 일치하지 않습니다.");
-    }
-}
