--- withins_server/api/src/main/java/com/withins/api/mail/repository/MailSendRepository.java
@@ -1,11 +1,0 @@
-package com.withins.api.mail.repository;
-
-import com.withins.api.mail.dto.MailSendDto;
-
-
-public interface MailSendRepository {
-
-    void sendTextEmail(MailSendDto mailSendDto);
-    void sendHtmlEmail(MailSendDto mailSendDto);
-
-}
