+++ withins_server/mail/src/main/java/com/withins/mail/exceptions/EmailConnectionException.java
@@ -1,0 +1,16 @@
+package com.withins.mail.exceptions;
+
+import lombok.Getter;
+
+@Getter
+public class EmailConnectionException extends EmailInternalException {
+
+    private final String host;
+    private final int port;
+
+    public EmailConnectionException(String host, int port) {
+        super("SMTP 서버 주소 또는 포트가 올바르지 않습니다.");
+        this.host = host;
+        this.port = port;
+    }
+}
