+++ withins_server/mail/src/main/java/com/withins/mail/exceptions/MailTemplateNotLoadedException.java
@@ -1,0 +1,28 @@
+package com.withins.mail.exceptions;
+
+import com.withins.mail.enums.EmailTemplate;
+import lombok.Getter;
+
+import java.util.Collections;
+import java.util.List;
+
+@Getter
+public class MailTemplateNotLoadedException extends EmailInternalException {
+
+    private final List<String> notFountTemplateList;
+
+    public MailTemplateNotLoadedException(List<String> resourceNames) {
+        super("[" + EmailTemplate.class.getName() + "] 이메일 템플릿을 로드하지 못했습니다. 찾을 수 없는 파일 : " + resourceNames);
+        this.notFountTemplateList = resourceNames;
+    }
+
+    public MailTemplateNotLoadedException(String resourceName) {
+        super("[" + EmailTemplate.class.getName() + "] 이메일 템플릿을 로드하지 못했습니다. 찾을 수 없는 파일 : " + resourceName);
+        this.notFountTemplateList = List.of(resourceName);
+    }
+
+    public MailTemplateNotLoadedException() {
+        super("[" + EmailTemplate.class.getName() + "] 알 수 없는 이유로 이메일 템플릿을 로드하지 못했습니다.");
+        this.notFountTemplateList = Collections.emptyList();
+    }
+}
