+++ withins_server/mail/src/main/java/com/withins/mail/repository/MailSendRepository.java
@@ -1,0 +1,11 @@
+package com.withins.mail.repository;
+
+import com.withins.mail.dto.MailSendDto;
+
+
+public interface MailSendRepository {
+
+    void sendTextEmail(MailSendDto mailSendDto);
+    void sendHtmlEmail(MailSendDto mailSendDto);
+
+}
